gedit: Better escape-sequence matching
Rather than matching /\\./ we now match proper lua escape
sequeces when in lua strings, or conky escape sequences when in
lua literal strings.

We also match conky template escape sequences when in lua literal
strings as it would be too messy to seperate `conky.text` and
`templateN = ` strings.

We currently don't match time format escape sequences as this would
require knowing which text variable we are in. This is only possible
if we duplicate the `bracket-var` context (less painful than
duplicating the `lua-literal-string` context).